refactor: add response dtos
