+++ src/modules/product/dto/filter-products.response.dto.ts
@@ -1,0 +1,11 @@
+import { ApiProperty } from '@nestjs/swagger';
+
+import { ProductEntity } from 'src/entities/product.entity';
+
+export class ProductsFilterResponseDto {
+  @ApiProperty()
+  products: ProductEntity[];
+
+  @ApiProperty()
+  totalCount: number;
+}
