+++ src/modules/product/product.controller.ts
@@ -1,18 +1,10 @@
-import {
-  Controller,
-  Get,
-  HttpCode,
-  HttpStatus,
-  Query,
-  //  UseGuards,
-} from '@nestjs/common';
+import { Controller, Get, HttpCode, HttpStatus, Query } from '@nestjs/common';
 import { ApiBearerAuth, ApiTags } from '@nestjs/swagger';
 
 import { ERouteName } from 'src/common/enums/route-name.enum';
-import { ProductsFilterDto } from 'src/modules/product/dto/filter-products.dto';
-// import { JwtAuthGuard } from 'src/modules/auth/guards/jwtAuthGuard';
+import { ProductsFilterDto } from 'src/modules/product/dto/filter-products.request.dto';
+import { ProductsFilterResponseDto } from 'src/modules/product/dto/filter-products.response.dto';
 import { ProductService } from 'src/modules/product/services/product.service';
-import { GetAllProductsOutput } from 'src/modules/product/types/get-all-products.type';
 
 @ApiTags('Product')
 @Controller(ERouteName.PRODUCTS_ROUTE)
@@ -24,7 +16,7 @@ export class ProductController {
   @HttpCode(HttpStatus.OK)
   async getAllProducts(
     @Query() query: ProductsFilterDto,
-  ): Promise<GetAllProductsOutput> {
+  ): Promise<ProductsFilterResponseDto> {
     return await this.productService.getAllProducts(query);
   }
 }
