+++ src/modules/product/services/product.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 
-import { ProductsFilterDto } from 'src/modules/product/dto/filter-products.dto';
-import { GetAllProductsOutput } from 'src/modules/product/types/get-all-products.type';
+import { ProductsFilterDto } from 'src/modules/product/dto/filter-products.request.dto';
+import { ProductsFilterResponseDto } from 'src/modules/product/dto/filter-products.response.dto';
 import { ProductRepository } from 'src/modules/repository/services/product.repository';
 
 @Injectable()
@@ -10,7 +10,7 @@ export class ProductService {
 
   public async getAllProducts(
     dto: ProductsFilterDto,
-  ): Promise<GetAllProductsOutput> {
+  ): Promise<ProductsFilterResponseDto> {
     return await this.productRepository.getAllProducts(dto);
   }
 }
