--- src/modules/product/types/get-all-products.type.ts
@@ -1,6 +1,0 @@
-import { ProductEntity } from 'src/entities/product.entity';
-
-export type GetAllProductsOutput = {
-  totalCount: number;
-  products: ProductEntity[];
-};
