+++ src/modules/user/dto/filter-users.response.dto.ts
@@ -1,0 +1,10 @@
+import { ApiProperty } from '@nestjs/swagger';
+
+import { UserEntity } from 'src/entities/user.entity';
+
+export class UsersFilterResponseDto {
+  @ApiProperty()
+  users: UserEntity[];
+  @ApiProperty()
+  totalCount: number;
+}
