--- src/modules/user/types/get-all-users.type.ts
@@ -1,3 +1,0 @@
-import { UserEntity } from 'src/entities/user.entity';
-
-export type GetAllUsersOutput = { users: UserEntity[]; totalCount: number };
