+++ src/modules/user/user.controller.ts
@@ -23,7 +23,8 @@ import { ERouteName } from 'src/common/enums/route-name.enum';
 import { ReviewEntity } from 'src/entities/review.entity';
 import { UserEntity } from 'src/entities/user.entity';
 import { JwtAuthGuard } from 'src/modules/auth/guards/jwtAuthGuard';
-import { UsersFilterDto } from 'src/modules/user/dto/filter-users.dto';
+import { UsersFilterDto } from 'src/modules/user/dto/filter-users.request.dto';
+import { UsersFilterResponseDto } from 'src/modules/user/dto/filter-users.response.dto';
 import { MakeReviewRequestDto } from 'src/modules/user/dto/make-review.request.dto';
 import { RateRequestDto } from 'src/modules/user/dto/rate-request.dto';
 import { UpdateUserDto } from 'src/modules/user/dto/update-user.dto';
@@ -31,7 +32,6 @@ import { FollowService } from 'src/modules/user/services/follow.service';
 import { RateService } from 'src/modules/user/services/rate.service';
 import { ReviewService } from 'src/modules/user/services/review.service';
 import { UserService } from 'src/modules/user/services/user.service';
-import { GetAllUsersOutput } from 'src/modules/user/types/get-all-users.type';
 
 @ApiTags('User')
 @Controller(ERouteName.USERS_ROUTE)
@@ -62,7 +62,9 @@ export class UserController {
 
   @Get(ERouteName.GET_USERS_ROUTE)
   @HttpCode(HttpStatus.OK)
-  async getAll(@Query() query: UsersFilterDto): Promise<GetAllUsersOutput> {
+  async getAll(
+    @Query() query: UsersFilterDto,
+  ): Promise<UsersFilterResponseDto> {
     return await this.userService.getAllUsers(query);
   }
